[ADD] ev_workshop:  Added Business Logic and Applied view enhancements
Applied Business logic the workshop model on button action, which will change
the maintenance stage according to the boolean type of need_mechanic and
moreover to it, Implemented the total stock price of the particular brand to be
computed as the variants are added. Added an icon of the module through web_icon
in the menuitem section.